Implement player health and take damage
Set a default health of 5 for the player (public variable set in editor).
all enemies do 1 damage when attacking.
implemented a coroutine to have player flash red when damaged.
When player's health is <= 0, they are dead and the sprite disappears (lol).
added two more player states (stunned and dead) but they don't do anything yet.
when in either of the above states player can't move