Create dynamically updating health UI
With help from AI, created an image pool class and a Player Health UI class. The Player Health UI class has a function that dynamically draws "Max Health" number of hearts on screen on game start. These are currently anchored in the top right corner of the screen. When the player takes damage they are removed one by one.