Fixes bug when resetting lastPlayedNotes
Instead of clearing the queue, which causes Dequeue calls to fail and interrupt the function (because there's nothing left to dequeue), we reset the queue to contain all null values. This achieves the desired effect of resetting the queue in between instrument sessions, so that a melody needs to be played in one session.