Upgrade System
Finished up the dialogue choice system. It essentially pops up a new canvas with dialogue options Yes and No. They have pretty barebones gray buttons that turn blue when selected. We can add sound as well as an indication frame if we want.

Then, I removed player dash and locked it behind the statue. You can also get it by talking to the grave just left of the tree as a test. Options to add sound and a screen flash effect (which I think is pretty cool).

HUGE BUG: statue only asks question on the 2nd instance of the player interacting with it. I think it is a frame timing issue.

Also note that it won't ask Yes/No if you already have the upgrade. Sometimes the upgrade gets saved between resets (like when you stop and restart the game)... not sure what is going on there. Going to work on the save system next, so hopefully I figure it out.