AOE Attack
more powerful AOE attack is working, however I need to fix the cooldown time. Right now, the cooldown prevents you from doing ANY attack, not just the AOE attack. To prevent this for now I made the cooldown very short, but this will lead to unbalanced combat so needs to be fixed