undo changes that caused dialogue bug
i guess this caused dialogue bug! but warning, if a sign controller has no name and "has dialogue on melody" is true, the player gets stuck in dialogue state. so make sure every sign controller that has that marked as true also has a name and that it gets handled in the sign controller script