Fixed Dialogue Typewriter Effect
Changed character by character in TypeLine function to token by token. Treated input tag as a single token, as well as each character of a normal string.

I also adjusted the rigid body of the ice obstacle a bit and added dynamic sorting so the player could get behind it. Not sure if that was the best way (as no other objects have the old dynamic sorting script...) but it worked and didn't seem to cause any bugs.