Uploaded Reset Settings Button Script
Gotta clock out for today, but this is some legwork for adding a button to the settings menu that says "Reset to Default." I just don't know about buttons yet or UI controls (specifically how to get each button to appear selectable w/ arrow keys). Feel free to implement this if you have time, if not it will be a nice to have for me to get to.